Remove http status code from previous responses cache
This is creating a bug where we can only set placeholder variables for previous responses with a 200 status.

To solve this, I considered adding an additonal step:

`When set the variable {string} using the json path {string} from the previous {string} to {string} with status {int}`

But decided that currently it doesn't really see necessary to use the status to cache a response. So I thought it simpler to just
omit status from the cache key.